requirements.txt
# Procesamiento de PDFs
pymupdf==1.23.8  # Alias de PyMuPDF
pdfplumber==0.10.3  # Backup para procesamiento de PDFs
cryptography>=41.0.7  # Requerido por pdfplumber

# Búsqueda vectorial
faiss-cpu==1.7.4
numpy==1.24.3

# Embeddings y transformers
sentence-transformers==2.5.1
huggingface-hub==0.19.4
transformers==4.36.2
torch==2.1.0

# LLM
ctransformers==0.2.27

# Interfaz web
streamlit==1.31.1

# Utilidades
tqdm==4.66.1
pandas==2.2.3
scikit-learn==1.3.0
scipy==1.15.2

# LLM con soporte CUDA
ctransformers[cuda]==0.2.27 